Moved Horizontal barchart code from experimental module to Y chartLib.
