+++ YChartsLib/src/main/java/co/yml/charts/axis/AxisData.kt
@@ -44,6 +44,7 @@ data class AxisData(
     val axisTopPadding: Dp,
     val axisBottomPadding: Dp,
     val axisStartPadding: Dp,
+    val axisEndPadding: Dp,
     val axisStepSize: Dp,
     val axisLabelAngle: Float,
     val axisLineColor: Color,
@@ -56,7 +57,8 @@ data class AxisData(
     val axisConfig: AxisConfig,
     val startDrawPadding: Dp,
     val shouldDrawAxisLineTillEnd: Boolean,
-    val axisLabelDescription: (String) -> String
+    val axisLabelDescription: (String) -> String,
+    val dataCategoryOptions: DataCategoryOptions
 ) {
     class Builder {
         private var steps: Int = 1
@@ -66,6 +68,7 @@ data class AxisData(
         private var axisStartPadding: Dp = 10.dp
         private var axisOffset: Dp = 10.dp
         private var axisTopPadding: Dp = 20.dp
+        private var axisEndPadding: Dp = 20.dp
         private var axisBottomPadding: Dp = 10.dp
         private var axisStepSize: Dp = 30.dp
         private var axisConfig = AxisConfig()
@@ -80,6 +83,7 @@ data class AxisData(
         private var axisLabelAngle: Float = 0f
         private var shouldDrawAxisLineTillEnd: Boolean = false
         private var axisLabelDescription: (String) -> String = { label -> "X Axis label $label" }
+        private var dataCategoryOptions: DataCategoryOptions = DataCategoryOptions()
 
         fun steps(count: Int) = apply { this.steps = count }
 
@@ -102,6 +106,8 @@ data class AxisData(
         fun topPadding(padding: Dp) = apply { this.axisTopPadding = padding }
 
         fun startPadding(padding: Dp) = apply { this.axisStartPadding = padding }
+
+        fun endPadding(padding: Dp) = apply { this.axisEndPadding = padding }
 
         fun bottomPadding(padding: Dp) = apply { this.axisBottomPadding = padding }
 
@@ -126,8 +132,12 @@ data class AxisData(
         fun axisLabelDescription(description: (String) -> String) =
             apply { this.axisLabelDescription = description }
 
+        fun setDataCategoryOptions(dataCategoryOptions: DataCategoryOptions) = apply {
+            this.dataCategoryOptions = dataCategoryOptions
+        }
 
-        fun build() = co.yml.charts.axis.AxisData(
+
+        fun build() = AxisData(
             steps,
             labelData,
             axisPos,
@@ -136,6 +146,7 @@ data class AxisData(
             axisTopPadding,
             axisBottomPadding,
             axisStartPadding,
+            axisEndPadding,
             axisStepSize,
             axisLabelAngle,
             axisLineColor,
@@ -148,7 +159,8 @@ data class AxisData(
             axisConfig,
             startDrawPadding,
             shouldDrawAxisLineTillEnd,
-            axisLabelDescription
+            axisLabelDescription,
+            dataCategoryOptions
         )
     }
 }
@@ -160,11 +172,15 @@ data class AxisData(
  * @param shouldEllipsizeAxisLabel : true if should ellipsize the axis label at end  else false
  * @param minTextWidthToEllipsize : minimum width of the axis label post which label will be ellipsized
  * @param ellipsizeAt : position at which the label will be truncated or ellipsized
-
  */
 data class AxisConfig(
     val isAxisLineRequired: Boolean = true,
     val shouldEllipsizeAxisLabel: Boolean = false,
     val minTextWidthToEllipsize: Dp = 40.dp,
     val ellipsizeAt: TextUtils.TruncateAt = TextUtils.TruncateAt.END
+)
+
+data class DataCategoryOptions(
+    val isDataCategoryInYAxis: Boolean = false,
+    val isDataCategoryStartFromBottom: Boolean = false
 )
