+++ YChartsLib/src/main/java/co/yml/charts/common/extensions/Extensions.kt
@@ -249,3 +249,71 @@ internal fun Context.collectIsTalkbackEnabledAsState(): State<Boolean> {
 }
 
 private const val TALKBACK_PACKAGE_NAME = "com.google.android.marvin.talkback"
+
+/**
+return the shape that is used to mask a particular area for given top & bottom
+ */
+internal class ColumnClip(
+    private val leftPadding: Float,
+    private val topPadding: Float = 0f,
+    private val rightPadding: Float,
+    private val bottomPadding: Float
+) : Shape {
+    override fun createOutline(
+        size: Size,
+        layoutDirection: LayoutDirection,
+        density: Density
+    ): Outline {
+        return Outline.Rectangle(
+            androidx.compose.ui.geometry.Rect(
+                0f,
+                topPadding,
+                rightPadding,
+                size.height - bottomPadding
+            )
+        )
+    }
+}
+
+fun getMaxElementInXAxis(xMax: Float, xStepSize: Int): Int {
+    var reqYLabelsQuo =
+        (xMax / xStepSize)
+    val reqYLabelsRem = xMax.rem(xStepSize)
+    if (reqYLabelsRem > 0f) {
+        reqYLabelsQuo += 1
+    }
+    return reqYLabelsQuo.toInt() * xStepSize
+}
+
+/**
+ * Return true if the point is selected
+ * @param tapOffset Tapped offset
+ * @param yOffset in the Y axis
+ * @param left left Value
+ */
+fun Offset.isYAxisTapped(tapOffset: Offset, yOffset: Float, left: Float, tapPadding: Float) =
+    (((tapOffset.y) < y + (yOffset + tapPadding) / 2) && ((tapOffset.y) > y - (yOffset + tapPadding) / 2) &&
+            ((tapOffset.plus(Offset(tapPadding, 0f))).x > x) && ((tapOffset.x) > left))
+
+/**
+ * Returns the background rect for the highlighted text.
+ * @param x : X point.
+ * @param y: Y point.
+ * @param text: Text to be drawn inside the background.
+ * @param paint: Background paint.
+ */
+fun getYaxisTextBackgroundRect(
+    x: Float,
+    y: Float,
+    text: String,
+    paint: TextPaint
+): Rect {
+    val fontMetrics = paint.fontMetrics
+    val textLength = paint.measureText(text)
+    return Rect(
+        (x - (textLength / 2)).toInt(),
+        (y + fontMetrics.top).toInt(),
+        (x + (textLength / 2)).toInt(),
+        (y + fontMetrics.bottom).toInt()
+    )
+}
