+++ YChartsLib/src/main/java/co/yml/charts/ui/barchart/models/SelectionHighlightData.kt
@@ -16,6 +16,7 @@ import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import co.yml.charts.common.extensions.getTextBackgroundRect
+import co.yml.charts.common.extensions.getYaxisTextBackgroundRect
 
 /**
  * Used to customise the highlighted text and the bar
@@ -68,7 +69,7 @@ data class SelectionHighlightData(
         "$xLabel $yLabel"
     },
 
-    val drawPopUp: DrawScope.(Offset, BarData, Float) -> Unit = { selectedOffset, identifiedPoint, centerPointOfBar ->
+    val drawPopUp: DrawScope.(Offset, BarData, Float, Float, BarChartType) -> Unit = { selectedOffset, identifiedPoint, centerPointOfBar, selectedXAxisWidth, barChartType ->
         val highlightTextPaint = TextPaint().apply {
             textSize = highlightTextSize.toPx()
             color = highlightTextColor.toArgb()
@@ -77,9 +78,13 @@ data class SelectionHighlightData(
         }
         val label = popUpLabel(identifiedPoint.point.x, identifiedPoint.point.y)
         drawContext.canvas.nativeCanvas.apply {
-            val background = getTextBackgroundRect(
-                centerPointOfBar,
-                selectedOffset.y,
+            val x =
+                if (barChartType == BarChartType.VERTICAL) centerPointOfBar else selectedXAxisWidth + highlightTextOffset.toPx()
+            val y =
+                if (barChartType == BarChartType.VERTICAL) selectedOffset.y else centerPointOfBar
+            val background = getYaxisTextBackgroundRect(
+                x,
+                y,
                 label,
                 highlightTextPaint
             )
@@ -87,7 +92,7 @@ data class SelectionHighlightData(
                 color = highlightTextBackgroundColor,
                 topLeft = Offset(
                     background.left.toFloat(),
-                    background.top.toFloat() - highlightTextOffset.toPx()
+                    if (barChartType == BarChartType.VERTICAL) background.top.toFloat() - highlightTextOffset.toPx() else background.top.toFloat()
                 ),
                 size = Size(background.width().toFloat(), background.height().toFloat()),
                 alpha = highlightTextBackgroundAlpha,
@@ -96,20 +101,26 @@ data class SelectionHighlightData(
                 blendMode = backgroundBlendMode,
                 style = backgroundStyle
             )
+            val drawTextX =
+                if (barChartType == BarChartType.VERTICAL) centerPointOfBar else selectedXAxisWidth + highlightTextOffset.toPx()
+            val drawTextY =
+                if (barChartType == BarChartType.VERTICAL) selectedOffset.y - highlightTextOffset.toPx() else centerPointOfBar
             drawText(
                 label,
-                centerPointOfBar,
-                selectedOffset.y - highlightTextOffset.toPx(),
+                drawTextX,
+                drawTextY,
                 highlightTextPaint
             )
         }
     },
 
-    val drawHighlightBar: DrawScope.(Float, Float, Float, Float) -> Unit = { x, y, width, height ->
+    val drawHighlightBar: DrawScope.(Float, Float, Float, Float, BarChartType) -> Unit = { x, y, width, height, barChartType ->
+        val rectWidth = if (barChartType == BarChartType.VERTICAL) width else height
+        val rectHeight = if (barChartType == BarChartType.VERTICAL) height else width
         drawRoundRect(
             color = highlightBarColor,
             topLeft = Offset(x, y),
-            size = Size(width, height),
+            size = Size(rectWidth, rectHeight),
             cornerRadius = CornerRadius(
                 highlightBarCornerRadius.toPx(),
                 highlightBarCornerRadius.toPx()
@@ -161,5 +172,6 @@ data class SelectionHighlightData(
                 highlightTextPaint
             )
         }
-    }
+    },
+    val barChartType: BarChartType = BarChartType.VERTICAL
 )
