+++ app/src/main/java/co/yml/ycharts/app/presentation/BarChartActivity.kt
@@ -15,15 +15,16 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import co.yml.ycharts.app.ui.compositions.AppBarWithBackButton
-import co.yml.ycharts.app.ui.theme.YChartsTheme
 import co.yml.charts.axis.AxisData
+import co.yml.charts.common.utils.DataUtils
 import co.yml.charts.ui.barchart.BarChart
 import co.yml.charts.ui.barchart.models.BarChartData
+import co.yml.charts.ui.barchart.models.BarChartType
 import co.yml.charts.ui.barchart.models.BarStyle
 import co.yml.charts.ui.barchart.models.SelectionHighlightData
-import co.yml.charts.common.utils.DataUtils
 import co.yml.ycharts.app.R
+import co.yml.ycharts.app.ui.compositions.AppBarWithBackButton
+import co.yml.ycharts.app.ui.theme.YChartsTheme
 
 class BarChartActivity : ComponentActivity() {
 
@@ -65,7 +66,7 @@ class BarChartActivity : ComponentActivity() {
 @Composable
 private fun BarChart1() {
     val maxRange = 50
-    val barData = DataUtils.getBarChartData(50, maxRange)
+    val barData = DataUtils.getBarChartData(50, maxRange, BarChartType.VERTICAL)
     val yStepSize = 10
 
     val xAxisData = AxisData.Builder()
